% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOperations.R
\name{adj2el}
\alias{adj2el}
\title{Maps adjacency matrix to edgelist}
\usage{
adj2el(adj, directed = TRUE, names = NULL)
}
\arguments{
\item{adj}{matrix or sparse matrix object, the adjacency matrix}

\item{directed}{boolean, is the graph directed?}

\item{names}{optional vector with node names. When `NULL` row- and column
names of `adj` are used.}
}
\value{
a dataframe containing the edgelist
}
\description{
Maps adjacency matrix to edgelist
}
\examples{
el <- data.frame(from= c('a','b','b','c','d','d'),
                to  = c('b','c','d','a','b','a'),
                attr= c( 12, 6, 12 , 6 , 6 , 6 ))
adj <- el2adj(el)
el1 <- adj2el(adj, directed = TRUE)
}
