% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astibble_sparseM.R
\name{as_tibble.dgTMatrix}
\alias{as_tibble.dgTMatrix}
\alias{as_tibble.dgCMatrix}
\alias{as_tibble.sparseMatrix}
\title{Method to convert `Matrix` sparse matrices to `tibble`s
Method for `as_tibble` to coerce sparse matrix objects
to a `tibble`}
\usage{
\method{as_tibble}{dgTMatrix}(x, ...)

\method{as_tibble}{dgCMatrix}(x, ...)

\method{as_tibble}{sparseMatrix}(x, ...)
}
\arguments{
\item{x}{A Matrix object}

\item{...}{Extra arguments, not used}
}
\description{
Method to convert `Matrix` sparse matrices to `tibble`s
Method for `as_tibble` to coerce sparse matrix objects
to a `tibble`
}
