% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_specs.R
\name{check_specs}
\alias{check_specs}
\alias{check_specs.matrix}
\alias{check_specs.dgTMatrix}
\alias{check_specs.dgCMatrix}
\alias{check_specs.sparseMatrix}
\alias{check_specs.default}
\title{Check an adjacency matrix object for directedness and presence of selfloops}
\usage{
check_specs(object, ...)

\method{check_specs}{matrix}(object, ...)

\method{check_specs}{dgTMatrix}(object, ...)

\method{check_specs}{dgCMatrix}(object, ...)

\method{check_specs}{sparseMatrix}(object, ...)

\method{check_specs}{default}(object, ...)
}
\arguments{
\item{object}{matrix to be analyzed. Different classes are allowed.}

\item{...}{additional parameters to and from internal functions, currently
not used.}
}
\value{
named boolean vector of length 2 whose first element is TRUE for
  directed `objects` and whose second element is TRUE for `object` with
  selfloops.
}
\description{
This method allows different adjacency matrix classes to be checked.
}
\examples{
el <- data.frame(from= c('a','b','b','c','d','d'),
                to  = c('b','c','d','a','b','a'),
                attr= c( 12, 6, 12 , 6 , 6 , 6 ))
adj <- el2adj(el)
check_specs(adj)
}
