% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeconversion.R
\name{datestring_to_unix}
\alias{datestring_to_unix}
\title{Converts the given datestring into a unix-timestamp.}
\usage{
datestring_to_unix(datestring, in_format = "\%Y-\%m-\%d")
}
\arguments{
\item{datestring}{The date to convert given as a string of the specified
in_format.}

\item{in_format}{The format code of the given datestring. For the
specification see section "Details" in help(strptime)}
}
\value{
Integer representing the date as a unix_timestamp.
}
\description{
Converts the given datestring into a unix-timestamp.
}
\examples{
      datestring_to_unix("1970-01-01")
      # 0
      datestring_to_unix("1970-01-02")
      # 86400
}
