% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeconversion.R
\name{detect_date_format}
\alias{detect_date_format}
\title{Detect date format of a string}
\usage{
detect_date_format(datestring)
}
\arguments{
\item{datestring}{string representing a date}
}
\value{
a string with the format code of the given datestring
}
\description{
Detect date format of a string
}
\examples{
detect_date_format("1970-01-01")

}
