% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_nets.R
\name{get_rolling_windows}
\alias{get_rolling_windows}
\title{Builds rolling multiedge time-window networks from the edge_list.}
\usage{
get_rolling_windows(
  el,
  window_size,
  step_size = NULL,
  start_time = NULL,
  end_time = NULL,
  out_format = "edgelist",
  flush = "earliest",
  select_cols = NULL,
  as_date = NULL,
  ncores = NULL,
  ...
)
}
\arguments{
\item{el}{A data-frame with at least 3 columns: "source", "target",
"timestamp", where "from" and "to" are vertex identifiers, and "time" is a
Unix-timestamp when the edge occurred. Alternatively, the user can select
which column is which passing `select_cols`}

\item{window_size}{The duration of each time-window network given as the
number of seconds.}

\item{step_size}{The shift between consecutive time-window network given as
the number of seconds.}

\item{start_time}{(optional) If given, the rolling time-windows are built
starting at this timestamp. Else, they are started at min(edgelist$time).}

\item{end_time}{(optional) If given, rolling time-windows are created until
they exceed this value. Else, this value defaults to max(edgelist$time).}

\item{out_format}{character vector specifying the format of the output:
'edgelist' for edgelists, 'adjacency' for adjacency matrices, `igraph` for
igraph graphs. Defaults to 'edgelist'. In the case of igraph graphs, adding
`igraph_weighted` allows to generated weighted graphs.}

\item{flush}{(optional) If "earliest" or unset, the first time-window starts
at `earliest`, possibly excluding an incomplete time-window at `latest`. If
"latest", the last time-window ends at `latest`, and possibly an incomplete
time-window at earliest is discarded.}

\item{select_cols}{an (optional) vector specifying which columns to use. The
first entry should contain either the name or index for the timestamp
column, the second for the sources of links, the third for the targets of
links. When not passed, it will be assumed that the first three colummns
are `c(timestamp,source,target)`.}

\item{as_date}{(optional) boolean identifying if timestamps are in date
formats, or unix seconds.}

\item{ncores}{integer, number of cores to use. Defaults to 1.}

\item{...}{additional parameters passed to internal constructors. E.g., to
`get_adjacency`.}
}
\value{
A tibble with two columns "network" and "time", where "network" is
  the time-window and "time" the corresponding time at which the time-window
  starts.
}
\description{
Builds rolling multiedge time-window networks from the edge_list.
}
\author{
CZ, GC
}
