% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_nets.R
\name{get_slice_edgelist}
\alias{get_slice_edgelist}
\title{Filter edgelist for a time slice}
\usage{
get_slice_edgelist(
  el,
  select_cols = NULL,
  start_time,
  end_time = NULL,
  duration = NULL,
  index = FALSE,
  as_date = NULL
)
}
\arguments{
\item{el}{edgelist}

\item{attr_cols}{}
}
\value{

}
\description{
Filter edgelist for a time slice
}
