% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixOperations.R
\name{mat2vec_ix}
\alias{mat2vec_ix}
\title{Auxiliary function, gives mask for matrix for directed,
undirected etc.}
\usage{
mat2vec_ix(x, directed, selfloops, arr.ind = FALSE)
}
\arguments{
\item{x}{matrix}

\item{directed}{a boolean argument specifying whether object is directed or not.}

\item{selfloops}{a boolean argument specifying whether the model should incorporate selfloops.}

\item{arr.ind}{boolean, return vector of indices or tuple with array indices? defaults to FALSE}
}
\value{
a boolean matrix that can be used to mask adjacency matrices.
}
\description{
Auxiliary function, gives mask for matrix for directed,
undirected etc.
}
\examples{
el <- data.frame(from= c('a','b','b','c','d','d'),
                to  = c('b','c','d','a','b','a'),
                attr= c( 12, 6, 12 , 6 , 6 , 6 ))
adj <- el2adj(el)
mat2vec_ix(adj, directed = TRUE, selfloops=FALSE)
}
