% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_maps.R
\name{node_to_id_map}
\alias{node_to_id_map}
\title{Map node names to their index in a nodes vector}
\usage{
node_to_id_map(x)
}
\arguments{
\item{x}{vector of names}
}
\value{
a named list
}
\description{
Generate a map from node names to their index in a nodes vector This function
return a named list whose elements are the indices of x named as the elements
of x.
}
\examples{
nodes <- c('a', 'b', 'c')
node_to_id <- node_to_id_map(nodes)
vec <- nodes[sample(3, 20, replace=TRUE)]
ids <- unlist(node_to_id[vec])
}
